Update 00_introduction.md
Dingtalk is not a commonly used office software overseas. It is advisable to consider Teams or other community methods. Before making a decision, it is recommended to remove it to avoid affecting the user experience. Even if it is determined to continue using DingTalk groups later, it is also necessary to consider creating separate groups or providing the DingTalk download channel at the same time.